Separate out types file
