+++ editor.planx.uk/src/pages/FlowEditor/ReadMePage/types.ts
@@ -1,0 +1,12 @@
+import { FlowInformation } from "../utils";
+
+export interface ReadMePageProps {
+  flowSlug: string;
+  flowInformation: FlowInformation;
+  teamSlug: string;
+}
+export interface ReadMePageForm {
+  serviceSummary: string;
+  serviceDescription: string;
+  serviceLimitations: string;
+}
